Split project into library and demo app
