+++ sensorfusionlib/src/main/java/org/hitlabnz/sensorfusionlib/orientationProvider/CalibratedGyroscopeProvider.java
@@ -1,6 +1,6 @@
-package org.hitlabnz.sensor_fusion_demo.orientationProvider;
+package org.hitlabnz.sensorfusionlib.orientationProvider;
 
-import org.hitlabnz.sensor_fusion_demo.representation.Quaternion;
+import org.hitlabnz.sensorfusionlib.representation.Quaternion;
 
 import android.hardware.Sensor;
 import android.hardware.SensorEvent;
