+++ sensorfusionlib/src/main/java/org/hitlabnz/sensorfusionlib/representation/Vector3f.java
@@ -1,4 +1,4 @@
-package org.hitlabnz.sensor_fusion_demo.representation;
+package org.hitlabnz.sensorfusionlib.representation;
 
 /**
  * 3-dimensional vector with conventient getters and setters. Additionally this class is serializable and
