+++ sensorfusionlib/src/main/java/org/hitlabnz/sensorfusionlib/representation/Vector4f.java
@@ -1,4 +1,4 @@
-package org.hitlabnz.sensor_fusion_demo.representation;
+package org.hitlabnz.sensorfusionlib.representation;
 
 /**
  * Representation of a four-dimensional float-vector
