+++ settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "SensorFusionDemo"
 include ':app'
+include ':sensorfusionlib'
